# MQE: encode query plan functions as an integer, rather than a string, and simplify function handling