--- a/pkg/streamingpromql/operators/functions/functions.go
+++ b/pkg/streamingpromql/operators/functions/functions.go
@@ -2,45 +2,8 @@
 
 package functions
 
-import (
-	"fmt"
-	"strings"
-)
-
-var functionsToPromQLNames = convertFunctionConstantsToPromQLNames()
-var promQLNamesToFunctions = invert(functionsToPromQLNames)
-
-func convertFunctionConstantsToPromQLNames() map[Function]string {
-	names := make(map[Function]string, len(Function_name))
-
-	for idx, constantName := range Function_name {
-		f := Function(idx)
-
-		if f == FUNCTION_UNKNOWN {
-			continue
-		}
-
-		names[f] = strings.ToLower(strings.TrimPrefix(constantName, "FUNCTION_"))
-	}
-
-	return names
-}
-
-func invert[A, B comparable](original map[A]B) map[B]A {
-	inverted := make(map[B]A, len(original))
-
-	for k, v := range original {
-		before := len(inverted)
-		inverted[v] = k
-		after := len(inverted)
-
-		if before == after {
-			panic(fmt.Sprintf("duplicate value %v detected", v))
-		}
-	}
-
-	return inverted
-}
+var functionsToPromQLNames = map[Function]string{}
+var promQLNamesToFunctions = map[string]Function{}
 
 func (f Function) PromQLName() string {
 	name, ok := functionsToPromQLNames[f]