--- a/pkg/streamingpromql/operators/functions/functions.go
+++ b/pkg/streamingpromql/operators/functions/functions.go
@@ -2,17 +2,16 @@
 
 package functions
 
-var functionsToPromQLNames = map[Function]string{}
 var promQLNamesToFunctions = map[string]Function{}
 
 func (f Function) PromQLName() string {
-	name, ok := functionsToPromQLNames[f]
+	fnc, ok := RegisteredFunctions[f]
 
 	if !ok {
 		return f.String()
 	}
 
-	return name
+	return fnc.Name
 }
 
 func FromPromQLName(name string) (Function, bool) {