… pkg/streamingpromql/operators/functions/factories_test.go
… pkg/streamingpromql/operators/functions/functions.pb.go
… pkg/streamingpromql/planning/core/core.pb.go
… pkg/streamingpromql/planning/core/core.proto
… pkg/streamingpromql/planning/core/function_call.go
… pkg/streamingpromql/planning/core/function_call_test.go
… pkg/streamingpromql/planning_test.go