#### What this PR does

This PR builds on #11936, and adopts a similar approach for functions.

This further reduces query plan size, and improves encoding and decoding performance slightly:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql
cpu: Apple M1 Pro
                                                                                                                               │ with-int.txt │     with-function-as-int.txt      │
                                                                                                                               │    sec/op    │   sec/op     vs base              │
PlanEncodingAndDecoding/sum(rate(foo[5m]))/encode-10                                                                              684.7n ± 0%   679.5n ± 0%  -0.75% (p=0.004 n=6)
PlanEncodingAndDecoding/sum(rate(foo[5m]))/decode-10                                                                              864.4n ± 0%   832.9n ± 0%  -3.64% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/encode-10    1.487µ ± 0%   1.468µ ± 0%  -1.24% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/decode-10    1.893µ ± 0%   1.853µ ± 0%  -2.09% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/encode-10                       2.769µ ± 0%   2.720µ ± 1%  -1.77% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/decode-10                       3.272µ ± 2%   3.185µ ± 0%  -2.64% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/encode-10                                       1.990µ ± 2%   1.970µ ± 0%  -0.98% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/decode-10                                       2.162µ ± 0%   2.113µ ± 2%  -2.24% (p=0.002 n=6)

                                                                                                                               │ with-int.txt │      with-function-as-int.txt      │
                                                                                                                               │      B       │     B       vs base                │
PlanEncodingAndDecoding/sum(rate(foo[5m]))/encode-10                                                                              100.00 ± 0%   96.00 ± 0%  -4.00% (p=0.002 n=6)
PlanEncodingAndDecoding/sum_by_(zone)_(label_replace(rate(foo[5m]),_"zone",_"$1",_"pod",_"ingester-zone-(.*)-\\d+"))/encode-10     301.0 ± 0%   284.0 ± 0%  -5.65% (p=0.002 n=6)
PlanEncodingAndDecoding/max_over_time(rate(foo[5m:])[10m:])_+_max_over_time(rate(bar[5m:])[10m:])/encode-10                        367.0 ± 0%   333.0 ± 0%  -9.26% (p=0.002 n=6)
PlanEncodingAndDecoding/label_join(foo,_"abc",_"-")_+_label_join(bar,_"def",_",")/encode-10                                        267.0 ± 0%   247.0 ± 0%  -7.49% (p=0.002 n=6)

```

The main drawback of this approach is that it makes handling additional functions (eg. those introduced by Grafana's Adaptive Metrics query rewriting) a little more awkward: we must either declare those functions in `functions.proto`, or at least reserve a value for them, to ensure that we don't end up with conflicting functions with the same value.

This PR also makes two other changes to how functions are handled:

* it removes the special cases for `absent` and `absent_over_time`
* it unifies the handling of functions that return scalars and instant vectors

I would recommend reviewing each commit separately.

#### Which issue(s) this PR fixes or relates to

#11936

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.